refactored endpoint, connection timeout
